Update to CraftBukkit 1.6.1

+++ src/com/gmail/scyntrus/fmob/FactionMob.java
@@ -1,8 +1,8 @@
 package com.gmail.scyntrus.fmob;
 
-import net.minecraft.server.v1_5_R3.Entity;
-import net.minecraft.server.v1_5_R3.EntityLiving;
-import net.minecraft.server.v1_5_R3.ItemStack;
+import net.minecraft.server.v1_6_R1.Entity;
+import net.minecraft.server.v1_6_R1.EntityLiving;
+import net.minecraft.server.v1_6_R1.ItemStack;
 
 import org.bukkit.Location;
 
@@ -19,8 +19,8 @@ public interface FactionMob {
 	public void updateMob();
 	public String getTypeName();
 	public boolean isAlive();
-	public int getHealth();
-	public void setHealth(int hp);
+	public float getHealth();
+	public void setHealth(float hp);
 	public void die();
 	public void setEquipment(int slot, ItemStack item);
 	public Boolean getEnabled();
